add home and room page
